functions/search/requirements.txt
parliament-functions==0.1.0
cloudevents==1.11.0
functions-framework==3.*
planetary-computer==1.0.0
pystac-client==0.8.6
minio==7.2.15
requests==2.32.3
werkzeug==2.2.3
gunicorn==20.1.0